Handle OpenID claims parameter in auth requests on Android
The claims parameter is an optional part of the OpenID standard when making an authorization request (see https://openid.net/specs/openid-connect-core-1_0.html#ClaimsParameter).

Currently, this parameter is not supported by the Flutter interface layer, so the only way to pass it is via the additionalParameters argument when constructing an AuthorizationRequest. This, however, leads to the following error on Android:

```
E/AndroidRuntime( 7453): java.lang.IllegalArgumentException: Parameter claims is directly supported via the authorization request builder, use the builder method instead
```

This commit allows specifying the claims via additionalParameters and having this be correctly handled on Android, in order to be compatible with the `AuthorizationRequest.Builder` method.