remove idTokenHint assert
I am wondering if there is a specific reason for this assert?

Currently we want to only supply our postLogoutRedirectUrl but not the idTokenHint in our app. We tried using the additionalParameters field to supply only the postLogoutRedirectUrl, but that causes fatal exception in android.

Would be nice to remove this assert, unless there is a reason with OpenID why we'd want to enforce it like this.

Let me know!
Thank you!